+++ lib/client/utilities/ct_supplier_backlog_rates.dart
@@ -24,18 +24,35 @@ enum Utility {
   String toString() => _value;
 }
 
-
 class CtSupplierBacklogRates {
   CtSupplierBacklogRates(http.Client client, {required this.rootUrl});
 
   final String rootUrl;
   final String servicePath = '/retail_suppliers/v1/ct/supplier_backlog_rates';
 
-  /// Get backlog data from all suppliers between two months.
-  Future<List<Map<String, dynamic>>> getBacklogForUtility(
-      {required Utility utility,
-      required Month start,
-      required Month end}) async {
+  /// Get backlog data from all suppliers between two months.  Return data
+  /// sorted by month, each element in this shape:
+  ///```
+  ///{
+  ///  'month': '2023-01',
+  ///  'utility': 'Eversource',
+  ///  'customerClass': 'C&I',
+  ///  'supplierName': 'CATALYST POWER & GAS LLC',
+  ///  'price': <num>[...],
+  ///  'kWh': <num>[...],
+  ///  'summary': {
+  ///    'customerCount': 97,
+  ///    'averagePriceWeightedByCustomerCount': 0.1424,
+  ///    'kWh': 529542.5,
+  ///    'averagePriceWeightedByVolume': 0.1419
+  ///  },
+  ///}
+  ///```
+  Future<List<Map<String, dynamic>>> getBacklogForUtility({
+    required Utility utility,
+    required Month start,
+    required Month end,
+  }) async {
     var url = '$rootUrl$servicePath/utility/${utility.toString()}'
         '/start/${start.toIso8601String()}'
         '/end/${end.toIso8601String()}';